refactor: Провести рефакторинг кода перед сдачей работы
