--- src/components/withFormValidation.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-function WithFormValidation(Component) {
-  console.log('hoc');
-  return <Component />;
-}
-
-export default WithFormValidation;
